changed include directories in cmake config
The previous version defines the include directory "include/DLib", where
there are no headers. On the other hand, "include" itself is not set as
include directory, which means the include files - when included with
subdir, i.e.  "DVision/DVision.h" - are only found if
"<install-prefix>/include" is already in the global search path for
includes, e.g., /usr/local/include.